# spec/openapi.yaml
...
openapi: 3.1.0
info:
  title: Ashurbanipal Protocol
  version: "1"
  description: >
    Read-only SQL table-browsing API (Postgres, MySQL, or SQLite
    backend). Paths are relative to
    {mount}/api, where {mount} is the implementation-defined mount point
    (reference default: /__ashurbanipal). The UI itself is served at
    {mount} as text/html and is not described here.
servers:
  - url: "{mount}/api"
    variables:
      mount:
        default: /__ashurbanipal
        description: Implementation-defined mount point.
paths:
...
  /schemas:
    get:
      operationId: listSchemas
      summary: List schema names selectable as the `schema` param elsewhere.
      parameters:
        - $ref: "#/components/parameters/Source"
      responses:
        "200":
          description: >
            Live allow-list for the `schema` param on /tables,
            /table-counts, /tables/data, /tables/common-values, and
            /tables/referenced-by (spec/protocol.md §5.7).
          headers:
            x-ashurbanipal-protocol:
              $ref: "#/components/headers/ProtocolVersion"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SchemasResponse"
        "400":
          $ref: "#/components/responses/ClientError"
        "500":
          $ref: "#/components/responses/ServerError"
  /tables:
    get:
      operationId: listTables
      summary: List base tables of the resolved schema.
      parameters:
        - $ref: "#/components/parameters/Source"
        - $ref: "#/components/parameters/Schema"
      responses:
        "200":
          description: Table list (also the allow-list for `table` params).
          headers:
            x-ashurbanipal-protocol:
              $ref: "#/components/headers/ProtocolVersion"
          content:
            application/json:
              schema:
                type: object
                required: [tables]
                properties:
                  tables:
                    type: array
                    items:
                      $ref: "#/components/schemas/TableInfo"
        "400":
          $ref: "#/components/responses/ClientError"
        "500":
          $ref: "#/components/responses/ServerError"
  /table-counts:
    get:
      operationId: tableCounts
      summary: Approximate row counts from pg_class.reltuples.
      parameters:
        - $ref: "#/components/parameters/Source"
        - $ref: "#/components/parameters/Schema"
      responses:
        "200":
          description: Approximate counts for every table.
          headers:
            x-ashurbanipal-protocol:
              $ref: "#/components/headers/ProtocolVersion"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CountsResponse"
        "400":
          $ref: "#/components/responses/ClientError"
        "500":
          $ref: "#/components/responses/ServerError"
...
  /tables/common-values:
    get:
      operationId: commonValues
      summary: Most-common values for one column, from pg_stats.
      parameters:
        - $ref: "#/components/parameters/Source"
        - $ref: "#/components/parameters/Schema"
        - name: table
          in: query
          required: true
          schema:
            type: string
          examples:
            # `feature_flags` is deliberately excluded from the seed's
            # ANALYZE pass (conformance/seed/README.md), so this pair
            # exercises the empty-values-not-an-error path
            # (spec/protocol.md §5.5) alongside the ordinary populated one.
            orders:
              value: orders
            feature_flags:
              value: feature_flags
        - name: column
          in: query
          required: true
          schema:
            type: string
          examples:
            orders:
              value: status
            feature_flags:
              value: enabled
      responses:
        "200":
          description: >
            Most-common values, most frequent first. Empty when the column
            has no planner statistics (never an error).
          headers:
            x-ashurbanipal-protocol:
              $ref: "#/components/headers/ProtocolVersion"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommonValuesResponse"
        "400":
          $ref: "#/components/responses/ClientError"
        "500":
          $ref: "#/components/responses/ServerError"
  /tables/referenced-by:
    get:
      operationId: referencedBy
      summary: Foreign keys elsewhere in the schema that target this table.
      parameters:
        - $ref: "#/components/parameters/Source"
        - $ref: "#/components/parameters/Schema"
        - name: table
          in: query
          required: true
          description: >
            Must exactly match (case-sensitive) a table name from /tables;
            otherwise 400.
          schema:
            type: string
          examples:
            # Three named cases from conformance/seed/seed.sql exercising
            # distinct 200 shapes (spec/protocol.md §5.9): a multi-referrer
            # target (users — two of the entries share a referencing
            # table), a composite-FK target (inventory_locations — one
            # entry whose `columns` has two pairs), and a target nothing
            # references (feature_flags — `referenced_by` is []).
            users:
              value: users
            inventory_locations:
              value: inventory_locations
            feature_flags:
              value: feature_flags
      responses:
        "200":
          description: >
            One entry per foreign-key constraint pointing at `table`;
            `referenced_by` is [] when nothing does (not an error).
          headers:
            x-ashurbanipal-protocol:
              $ref: "#/components/headers/ProtocolVersion"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReferencedByResponse"
        "400":
          $ref: "#/components/responses/ClientError"
        "500":
          $ref: "#/components/responses/ServerError"
...
components:
  parameters:
    Source:
      name: source
      in: query
      required: false
      description: >
        Selects which registered DbSource this operation targets
        (spec/protocol.md §1). Absent resolves to the first-registered
        source. Present must exactly match (case-sensitive) an entry
        from GET /sources; otherwise 400.
      schema:
        type: string
    Schema:
      name: schema
      in: query
      required: false
      description: >
        Selects which schema this operation is scoped to (spec/protocol.md
        §1). Absent resolves to the connection's own default. Present
        must exactly match (case-sensitive) an entry from GET /schemas;
        otherwise 400.
      schema:
        type: string
  headers:
    ProtocolVersion:
      description: Protocol version; "1" for this document.
      required: true
      schema:
        type: string
        const: "1"
  responses:
    ClientError:
      description: >
        Client error — unknown source/schema/table/column, invalid
        filter, invalid parameter, or a table the role can't read.
      headers:
        x-ashurbanipal-protocol:
          $ref: "#/components/headers/ProtocolVersion"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
    ServerError:
      description: Server error — database failure or query timeout.
      headers:
        x-ashurbanipal-protocol:
          $ref: "#/components/headers/ProtocolVersion"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
  schemas:
...
    SchemasResponse:
      type: object
      required: [schemas]
      properties:
        schemas:
          type: array
          items:
            type: string
          description: >
            Selectable schema names, excluding the engine's own
            system/internal namespaces. Exactly one entry on engines with
            no schema concept above a single database.
    TableInfo:
      type: object
      required: [name]
      properties:
        name:
          type: string
        comment:
          type: string
          description: COMMENT ON TABLE text; omitted (never null) when absent.
...
    CountsResponse:
      type: object
      required: [counts]
      properties:
        counts:
          type: array
          items:
            type: object
            required: [table, approx_rows]
            properties:
              table:
                type: string
              approx_rows:
                type: integer
                format: int64
                description: May be stale or -1 before first ANALYZE.
    CommonValuesResponse:
      type: object
      required: [values]
      properties:
        values:
          type: array
          description: Most frequent first; empty when no planner statistics.
          items:
            type: object
            required: [value, freq]
            properties:
              value:
                type: string
                description: >
                  Rendered consistently with row serialization (e.g.
                  booleans as "true"/"false").
              freq:
                type: number
                format: float
                description: Fraction of rows, in (0, 1].
    ReferencedByResponse:
      type: object
      required: [referenced_by]
      properties:
        referenced_by:
          type: array
          description: >
            One object per foreign-key constraint whose referenced table
            is the requested `table` (spec/protocol.md §5.9). A referencing
            table with two FKs to the target appears twice. Empty array
            when nothing references the target. SHOULD be ordered by
            (schema, table, constraint).
          items:
            type: object
            required: [table, constraint, columns]
            properties:
              table:
                type: string
                description: The referencing table (holds the FK). Bare name.
              schema:
                type: string
                description: >
                  The referencing table's schema; present only when it
                  differs from the resolved schema (a cross-schema
                  reference). This is the opposite end of the relationship
                  from ColumnInfo.references.schema.
              constraint:
                type: string
                description: >
                  FK constraint name; (table, constraint) is unique within
                  one response. MAY be a synthesized label on engines that
                  don't name FK constraints (e.g. SQLite) — not a DDL
                  identifier, not guaranteed stable.
              columns:
                type: array
                description: >
                  Column pairing, one entry per column in the constraint,
                  in the constraint's own order. More than one entry for a
                  composite FK (included here, unlike §5.4.1's per-column
                  metadata).
                items:
                  type: object
                  required: [from, to]
                  properties:
                    from:
                      type: string
                      description: Referencing column on `table`.
                    to:
                      type: string
                      description: Referenced column on the target table.
...
    ProblemDetails:
      type: object
      description: >
        RFC 9457 Problem Details (application/problem+json). Sent for
        every 4xx/5xx (spec/protocol.md §2). Additional members beyond
        those listed MAY appear; clients ignore unknown ones.
      required: [status, title, code]
      properties:
        type:
          type: string
          format: uri-reference
          default: about:blank
          description: >
            Ashurbanipal defines no problem-type URI namespace; this is
            about:blank in practice. `code` is the discriminator.
        title:
          type: string
          description: Short human-readable summary; implementation-defined, do not parse.
        status:
          type: integer
          description: HTTP status code, repeated in the body (RFC 9457 §3.1).
        detail:
          type: string
          description: Optional longer human-readable explanation; do not parse.
        instance:
          type: string
          format: uri-reference
          description: Optional; identifies the specific occurrence.
        code:
          type: string
          description: >
            Stable machine-readable problem token — the value clients
            branch on. The set is open (additive); an unrecognized value
            is treated as a generic error of the given `status`.
          enum:
            - unknown_source
            - unknown_schema
            - unknown_table
            - unknown_column
            - invalid_filter
            - invalid_parameter
            - not_readable
            - database_error
            - query_timeout